Switch to decimal rounded to 4 places for calculated numbers, for floating point precision.  Also added calculation unit tests

+++ EasyWorkout.Contracts/Responses/DataPointResponse.cs
@@ -1,4 +1,4 @@
 ﻿namespace EasyWorkout.Contracts.Responses
 {
-    public record DataPointResponse(Guid completedWorkoutId, DateTime completedDate, double totalVolume);
+    public record DataPointResponse(Guid completedWorkoutId, DateTime completedDate, decimal totalVolume);
 }
